refactor: move board to its own file
